% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{ages}
\alias{ages}
\title{Ages for all nodes}
\usage{
ages(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The age (time since root divergence) of all nodes in the phylogeny.
}
\description{
Ages for all nodes
}
