% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{ancestors}
\alias{ancestors}
\title{Ancestors for all nodes}
\usage{
ancestors(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
A list holding the node indices of the sequence of ancestors that lead
to each node.
}
\description{
Ancestors for all nodes
}
