% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{brlens}
\alias{brlens}
\title{Branch lengths for all edges}
\usage{
brlens(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The branch lengths of all edges in the phylogeny.
}
\description{
Branch lengths for all edges
}
