% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{children}
\alias{children}
\title{Immediate descendants of all nodes}
\usage{
children(phy, node)
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{node}{Index of node whose children are desired (may be ommitted).}
}
\value{
A list holding the node indices of the immediate descendants of
each node or an integer vector of node indices of the children of the given
node.
}
\description{
Immediate descendants of all nodes
}
