% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{descendants}
\alias{descendants}
\title{Return the descendants of a node}
\usage{
descendants(
  node,
  phy,
  visit = c("ALL_NODES", "INTERNAL_NODES_ONLY"),
  order = c("PREORDER", "POSTORDER")
)
}
\arguments{
\item{node}{Index of node whose descendants are desired.}

\item{phy}{An object of class \code{tree}.}

\item{visit}{If \code{ALL_NODES} returned descendants include terminal and
internal nodes. If \code{INTERNAL_NODES_ONLY} terminal taxa are omitted.}

\item{order}{Descendants may be returned in \code{PREORDER} or \code{POSTORDER}
sequence.}
}
\value{
A vector holding the node indices of the descendants of \code{node}
}
\description{
Return the descendants of a node
}
