% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{extract.clade}
\alias{extract.clade}
\title{Extract a clade}
\usage{
extract.clade(phy, node)
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{node}{Index of node whose subtree is desired.}
}
\value{
An object of class \code{tree} containing all nodes and edges
descended from \code{node}.
}
\description{
Extract a clade
}
