% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastaRNA}
\alias{fastaRNA}
\title{alignment of RNA sequences for the spike protein of some sarbecoviruses
closely related to SARSCoV2 in fasta format}
\format{
a matrix of class DNAbin
\describe{\item{attribute (dimnames)}{sequence name (accession number)}}
}
\source{
the \href{https://www.ncbi.nlm.nih.gov/genbank/}{genbank} database of
the National Center for Biotechnology Information or the Universal Protein
Resource \href{https://www.uniprot.org/}{UniProt} database
}
\usage{
fastaRNA
}
\description{
14 different sequences of some sarbecoviruses closely related to SARSCoV2
were retrieved from public databases and compared using CLUSTAL omega (1.2.4)
alignment software with the output in fasta format (see details for
the complete protocol for retrieving, saving and processing the data)
}
\details{
\enumerate{
\item spikeRNA_annot <- read.csv(system.file('extdata', 'spike_annot.csv',
package='firstPackage'))
\item accessions <- spikeRNA_annot$accession
\item download manually each sequence from 'genbank' and save as spike.fasta and
perform an alignment with clustal omega with following commands
(it needs to be installed locally or use a web interface)
\item clustalo --threads=4 -v --outfmt=fasta -t RNA -i
./data/sequences/spike.fasta -o ./data/sequences/spike_align.fasta
\item fastaRNA_load <- load_alignment(system.file('extdata', 'spike_align.fasta',
package='firstPackage'), 'fasta')
}
}
\keyword{datasets}
