% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_alignment}
\alias{load_alignment}
\alias{\alias{load_alignment}}
\title{function to load a file containing a sequences alignment generated by an
alignment software}
\usage{
load_alignment(file, format, type = "protein")
}
\arguments{
\item{file}{a character string specifying the path to the file containing a
sequences alignment to be loaded to the global environment}

\item{format}{a character string specifying the format of the file, i.e.
fasta, phylip, mase, clustal or msf (no default)}

\item{type}{a character string specifying type of sequences, i.e. protein,
DNA, RNA (default: protein)}
}
\value{
if the input type is DNA or RNA this function returns an object of
class DNAbin, if the input type is protein or unspecified (default) this
function returns an object of class alignment (seqinr package)
}
\description{
this function is wrapper around the functions
\code{'\link[seqinr]{read.alignment}'},
\code{'\link[adegenet]{fasta2DNAbin}'} and \code{'\link[ape]{as.DNAbin}'}.
The goal of this function is to load different types of sequences alignments,
i.e. protein, DNA, RNA, in different formats, i.e. fasta, phylip, mase,
clustal or msf and to convert it in the best suitable object class for
generating a distance matrix either with the 'seqnir' or the 'ape' package
}
\examples{
fastaRNA_load <- load_alignment(system.file('extdata',
                                            'spike_align.fasta',
                                            package='firstPackage'),
                                 'fasta', 'RNA')
phylipRNA_load  <- load_alignment(system.file('extdata',
                                              'spike_align.phy',
                                              package='firstPackage'),
                                 'phylip', 'RNA')
phylipProt_load <- load_alignment(system.file('extdata',
                                              'spike_prot_align.phy',
                                              package='firstPackage'),
                                 'phylip', 'protein')
clustalRNA_load <- load_alignment(system.file('extdata',
                                              'spike_align.aln',
                                              package='firstPackage'),
                                 'clustal')
msfRNA_load <- load_alignment(system.file('extdata',
                                          'orf1ab_align.msf',
                                          package='firstPackage'),
                                 'msf', 'RNA')
maseProtein_load <- load_alignment(system.file('extdata',
                                               'prot.mase',
                                               package='firstPackage'),
                                 'mase')
data(fastaRNA); stopifnot(identical(fastaRNA, fastaRNA_load))
data(phylipRNA); stopifnot(identical(phylipRNA, phylipRNA_load))
data(clustalRNA); stopifnot(identical(clustalRNA, clustalRNA_load))
data(msfRNA); stopifnot(identical(msfRNA, msfRNA_load))
data(maseProtein); stopifnot(identical(maseProtein, maseProtein_load))

}
\seealso{
\code{'\link{retrieve_seqs}'} \code{'\link{print_alignment}'}
\code{'\link{clean_alignment}'} \code{'\link{make_tree}'}
\code{'\link{max_parsimony}'}
}
\author{
gerardo esteban antonicelli
}
