% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max_parsimony}
\alias{max_parsimony}
\alias{\alias{max_parsimony}}
\title{function to calculate a maximum pasimony phylogenetic tree}
\usage{
max_parsimony(alignment, type, clustering = bionj, outgroup = NULL)
}
\arguments{
\item{alignment}{an object of class alignment or DNAbin containing a DNA, RNA
or protein sequences alignment}

\item{type}{a character string without '' specifying the type of sequences,
i.e. DNA, RNA or protein (without default value)}

\item{clustering}{a character string without '' specifying the clustering
algorithm to be used to build the phylogenetic tree, i.e classic
neighbor-joining (nj), improved neighbor-joining (bionj) (default), NJ or
bio-NJ from a distance matrix with possibly missing values (njs, bionjs),
balanced minimum evolution principle (fastme.bal), ordinary least-squares
minimum evolution principle (fastme.ols)}

\item{outgroup}{a character string without '' specifying a taxon to be taken
as root for building the tree. The default value is NULL leading to an
unrooted tree}
}
\value{
the function returns an object of class 'phylo' of the 'ape' package.
Advanced and elaborated plots can be drawn in later steps based on the tree
data of the phylo class object
}
\description{
this function is a wrapper around the functions
\code{'\link[seqinr]{dist.alignment}'}, \code{'\link[ape]{dist.dna}'},
\code{'\link[ape]{nj}'}, \code{'\link[ape]{bionj}'},
\code{'\link[ape]{fastme.bal}'}, \code{'\link[ape]{fastme.ols}'} and
\code{'\link[phangorn]{optim.parsimony}'}. it takes a sequences alignment in
format 'alignment' of 'DNAbin' matrix and perform all transformations and
steps to calculate a phylogenetic distance matrix based on similarity or
identity in the case of proteins or based in evolutionary models in the case
of DNA or RNA, to perform a phylogenetic clustering and to optimise the
phylogeny by a maximum parsimony algorithm
}
\examples{
data(fastaRNA)
data(phylipRNA)
data(phylipProt)
clustalRNA <- load_alignment(system.file('extdata',
                                         'spike_align.aln',
                                         package='firstPackage'),
                             'clustal')
fastaRNAtree <- max_parsimony(alignment=fastaRNA, type=RNA,
                              clustering=fastme.ols)
phylipRNAtree <- max_parsimony(phylipRNA, RNA, fastme.bal)
phylipProtTree <- max_parsimony(phylipProt, protein, outgroup=YP_0010399)
clustalRNAtree <- max_parsimony(clustalRNA, type=RNA, clustering=bionj)
\dontrun{plot.phylo(phylipRNAtree, type='u')}

}
\seealso{
\code{'\link{retrieve_seqs}'} \code{'\link{print_alignment}'}
\code{'\link{clean_alignment}'} \code{'\link{load_alignment}'}
\code{'\link{make_tree}'}
}
\author{
gerardo esteban antonicelli
}
