% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msfRNA}
\alias{msfRNA}
\title{alignment of RNA sequences for the replicase of some sarbecoviruses
closely related to SARSCoV2 in msf format}
\format{
a matrix of class DNAbin
\describe{\item{attribute (dimnames)}{sequence name (accession number)}}
}
\source{
the \href{https://www.ncbi.nlm.nih.gov/genbank/}{genbank} database of
the National Center for Biotechnology Information or the Universal Protein
Resource \href{https://www.uniprot.org/}{UniProt} database
}
\usage{
msfRNA
}
\description{
14 different sequences of some sarbecoviruses closely related to SARSCoV2
were retrieved from public databases and compared using CLUSTAL omega (1.2.4)
alignment software with the output in msf format (see details for
the complete protocol for retrieving, saving and processing the data)
}
\details{
\enumerate{
\item orf1abRNA_annot <- read.csv(system.file('extdata', 'orf1ab_annot.csv',
package='firstPackage'))
\item accessions <- orf1abRNA_annot$accession
\item download manually each sequence from 'genbank' and save as orf1ab.fasta and
perform an alignment with clustal omega with following commands
(it needs to be installed locally or use a web interface)
\item clustalo --threads=4 -v --outfmt=msf -t RNA -i
./data/sequences/orf1ab.fasta -o ./data/sequences/orf1ab_align.msf
\item msfRNA_load <- load_alignment(system.file('extdata', 'orf1ab_align.msf',
package='firstPackage'), 'msf', 'RNA')
}
}
\keyword{datasets}
