% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{Nnode}
\alias{Nnode}
\title{Number of nodes in a phylogeny}
\usage{
Nnode(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The number of terminal and internal nodes in the phylogeny.
}
\description{
Number of nodes in a phylogeny
}
