% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{notes}
\alias{notes}
\title{Notes for nodes embedded in a Newick strings}
\usage{
notes(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The embedded notes
}
\description{
Notes for nodes embedded in a Newick strings
}
