% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{Ntip}
\alias{Ntip}
\title{Number of terminal taxa in a phylogeny}
\usage{
Ntip(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The number of terminal taxa in the phylogeny
}
\description{
Number of terminal taxa in a phylogeny
}
