% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{parent}
\alias{parent}
\title{Immediate ancestor for all nodes}
\usage{
parent(phy, node)
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{node}{Index of node(s) whose parents are desired.}
}
\value{
An integer vector of immediate ancestors for the given nodes.
}
\description{
Immediate ancestor for all nodes
}
