% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylipProt}
\alias{phylipProt}
\title{alignment of protein sequences of the spike protein of some sarbecoviruses
closely related to SARSCoV2}
\format{
a list of class alignment with 4 elements
\describe{\item{nb}{number of sequences compared}
\item{nam}{sequence name (accession number)}
\item{seq}{a list of protein sequences}
\item{com}{NA}}
}
\source{
the \href{https://www.ncbi.nlm.nih.gov/genbank/}{genbank} database of
the National Center for Biotechnology Information or the Universal Protein
Resource \href{https://www.uniprot.org/}{UniProt} database
}
\usage{
phylipProt
}
\description{
14 different sequences of some sarbecoviruses closely related to SARSCoV2
were retrieved from public databases and compared using CLUSTAL omega (1.2.4)
alignment software with the output in phylip format (see examples section for
the complete protocol for retrieving, saving and processing the data)
}
\details{
\enumerate{
\item spikeProt_annot <- read.csv(system.file('extdata', 'spike_prot_annot.csv',
package='firstPackage'))
\item accessions <- spikeProt_annot$accession
\item download manually each sequence from 'UniProt' and save as spike_prot.fasta and
perform an alignment with clustal omega with following commands
(it needs to be installed locally or use a web interface)
\item clustalo --threads=4 -v --outfmt=phy -t Protein -i
./data/sequences/spike_prot.fasta -o ./data/sequences/spike_prot_align.phy
\item phylipProt_load <- load_alignment(system.file('extdata', 'spike_prot_align.phy',
package='firstPackage'), 'phylip', 'protein')
}
}
\keyword{datasets}
