\name{phylo}

\alias{phylo}

\docType{package}

\title{main: support for 'phylo' objects}

\description{
	\pkg{phylo} provides support for dealing with 'phylo' objects. 
}

\details{
 \tabular{ll}{
 Package: \tab phylo\cr
 Type: \tab Package\cr
 License: \tab GPL\cr
 LazyLoad: \tab yes\cr
 }
}

\author{
Jonathan M Eastman <jonathan.eastman@gmail.com>
}


