% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{plot.tree}
\alias{plot.tree}
\title{Plot a phylogeny}
\usage{
\method{plot}{tree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tree}}

\item{...}{Further arguments to control plot appearance}
}
\value{
Invisibly returns the plotting coordinates.
}
\description{
Plot a phylogeny
}
\details{
Additional arguments that can be specified through \code{...} include,
\describe{
\item{layout}{The coordinate system to use. Either "cartesian" or "polar".}
\item{direction}{One of "right", "left", "up", "down". Determines the direction
the terminal edges will point. Only valid if \code{layout = "cartesian"}.}
\item{lwd}{Line width for drawing edges. If the length is greater than 1 it
should be equal \code{Nnode(phy)}.}
\item{edge.color}{Color for drawing edges. If the length is greater than 1 it
should be equal \code{Nnode(phy)}.}
\item{tip.labels}{Either \code{TRUE} or \code{FALSE}. Specifies whether terminal
taxon labels are drawn.}
\item{cex.label}{The size to draw labels when \code{tip.labels = TRUE}.}
\item{offset}{Character offset to position labels when \code{tip.labels = TRUE}.}
\item{vtheta}{The angle between the first and last terminal node. Only valid
if \code{layout = "polar"}.}
\item{rbf}{The length the root branch as a fraction of tree height. Only valid
if \code{layout = "polar"}.}}
}
