% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{read.newick}
\alias{read.newick}
\title{Phylogenetic tree input}
\usage{
read.newick(file, text = NULL)
}
\arguments{
\item{file}{A filename pointing to a file containing a Newick character string.}

\item{text}{A Newick character string. If not \code{NULL} any \code{file}
argument is ignored.}
}
\value{
An object of class \code{tree}
}
\description{
Parse a phylogenetic tree in Newick string format
}
\seealso{
\code{\link{write.newick}}
}
