% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_seqs}
\alias{retrieve_seqs}
\alias{\alias{retrieve_seqs}}
\title{function to automatically retrieve sequences from a databank based on
accession numbers}
\usage{
retrieve_seqs(seqnames, acnucdb)
}
\arguments{
\item{seqnames}{a character vector specifying the accession numbers of the
sequences to be retrieved}

\item{acnucdb}{a character string specifying the name of the ACNUC database
to be searched. Use \code{'\link[seqinr]{choosebank}'} without arguments to
see a list of available databases}
}
\value{
this function returns a named list with the sequences matching the
sequence names of class SeqAcnucWeb retrieved by
\code{'\link[seqinr]{query}'} based on the provided accession numbers the
name of each list item is composed of the accession number and the length
of the sequence
}
\description{
this is a function to retrieve sequences of DNA, RNA or protein from
databases structured under ACNUC and located on the web, the retrieval is
limited to accession numbers as selection criteria, remote access to ACNUC
databases works by opening a socket connection on a port (for example on port
number 5558 at pbil.univ-lyon1.fr) and by communicating on this socket
following the protocol described
\href{http://doua.prabi.fr/databases/acnuc/remote_acnuc.html}{here}
}
\examples{
seqs <- retrieve_seqs(c('P06747', 'P0C569', 'O56773', 'Q5VKP1'), 'swissprot')
\dontrun{viruses_annot <- read.csv(system.file('extdata',
                                               'CoV_genomes_annot.csv',
                                               package='firstPackage'))}
\dontrun{accessions <- viruses_annot$accession}
\dontrun{seqs <- retrieve_seqs(accessions, 'genbank')}
\dontrun{str(seqs)}

}
\seealso{
\code{'\link{print_alignment}'} \code{'\link{clean_alignment}'}
\code{'\link{load_alignment}'} \code{'\link{make_tree}'}
\code{'\link{max_parsimony}'}
}
\author{
gerardo esteban antonicelli
}
