% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{rotate.node}
\alias{rotate.node}
\title{Rotate a node}
\usage{
rotate.node(node, phy)
}
\arguments{
\item{node}{A vector of node indices to rotate.}

\item{phy}{An object of class \code{tree}.}
}
\description{
Swaps the left and right descendants
}
