% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{tiplabels}
\alias{tiplabels}
\title{Terminal taxa labels}
\usage{
tiplabels(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\value{
The labels of the terminal taxa in the phylogeny.
}
\description{
Terminal taxa labels
}
