% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{tree.duplicate}
\alias{tree.duplicate}
\title{Create a copy of a phylogeny}
\usage{
tree.duplicate(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\description{
Because the R phylogeny object is just an external pointer to a
C structure any changes to the underlying C structure are propagated
to all R objects that point to that object. To avoid side-effects
of this type make a copy of the phylogeny first.
}
