% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{tree.isbinary}
\alias{tree.isbinary}
\title{Determine if a phylogeny is strictly bifurcating}
\usage{
tree.isbinary(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\description{
Determine if a phylogeny is strictly bifurcating
}
