% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{tree.isultrametric}
\alias{tree.isultrametric}
\title{Determine if a phylogeny is ultrametric}
\usage{
tree.isultrametric(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\description{
Determine if a phylogeny is ultrametric
}
