% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{tree.ladderize}
\alias{tree.ladderize}
\title{Ladderize a phylogeny}
\usage{
tree.ladderize(phy)
}
\arguments{
\item{phy}{An object of class \code{tree}.}
}
\description{
Rotates all nodes such that the descendant with the larger
subtree is placed on the left.
}
