% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{write.newick}
\alias{write.newick}
\title{Phylogenetic tree output}
\usage{
write.newick(phy, file = "")
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{file}{A filename pointing to a file into which the Newick string will
be written. If empty the output is written to the R console.}
}
\description{
Serialize a \code{tree} object to a Newick string
}
\seealso{
\code{\link{read.newick}}
}
